+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to the "cph-ng" extension will be documented in this file.
 ## [Unreleased]
 
 - **fix**: ensure document is saved before proceeding with compilation
+- **fix**: add message for compilation timeout
 
 ## 0.2.2